requirements.txt
opencv-python
numpy
PySDL2
scikit-image